backend/requirements.txt
amqp==2.5.1
anyjson==0.3.3
astroid==2.3.3
billiard==3.6.1.0
celery==4.3.0
Cerberus==1.3.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.0.1
django-appconf==1.0.3
django-celery-email==2.0.2
django-cors-headers==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.10.2
idna==2.8
importlib-metadata==0.23
isort==4.3.21
kombu==4.6.4
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
Pillow==7.2.0
pylint==2.4.4
python-decouple==3.1
pytz==2019.3
qrcode==6.1
raven==6.10.0
redis==3.3.8
requests==2.22.0
serpy==0.3.1
six==1.12.0
stripe==2.36.0
typed-ast==1.4.0
Unidecode==1.0.23
urllib3==1.25.3
vine==1.3.0
wrapt==1.11.2
zipp==0.6.0
